requirements.txt
flake8==7.0.0
flake8-quotes==3.4.0
Flask==3.0.2
Flask-WTF==1.2.1
h3==3.7.7
h3pandas==0.2.6
itsdangerous==2.1.2
Jinja2==3.1.3
keras==2.14.0;python_version>="3.11"
keras==2.12.0;python_version=="3.8"
matplotlib==3.8.3;python_version>="3.11"
matplotlib==3.7.5;python_version=="3.8"
numpy==1.26.4;python_version>="3.11"
numpy==1.23.5;python_version=="3.8"
pandas==2.1.2;python_version>="3.11"
pandas==2.0.3;python_version=="3.8"
pillow==10.2.0
pipenv==2023.12.1
prison==0.2.1
python-magic==0.4.27
PyYAML==6.0.1
scikit-learn==1.4.1.post1;python_version>="3.11"
scikit-learn==1.3.2;python_version=="3.8"
scipy==1.12.0;python_version>="3.11"
scipy==1.10.1;python_version=="3.8"
six==1.16.0
tensorflow==2.14.0;python_version>="3.11"
tensorflow==2.12.0;python_version=="3.8"
tifffile==2024.2.12;python_version>="3.11"
tifffile==2023.7.10;python_version=="3.8"
umap-learn==0.5.5
uWSGI==2.0.24
Werkzeug==3.0.1
WTForms==3.1.2
